requirements.txt
requests
feedparser
schedule
pytz
python-dotenv
